requirements.txt
tensorflow==2.13.0
numpy
opencv-python
matplotlib
scikit-learn
Pillow
streamlit
requests